Resizable Futoshiki puzzle game
Futoshiki is a board-based puzzle game, also known under the name
Unequal. It is playable on a square board having a given fixed size (4x4
for example). The purpose of the game is to discover the digits hidden
inside the board's cells; each cell is filled with a digit between 1 and
the board's size.
Learn more at: www.futoshiki.org